Fix UI toggle cell logic to work with json-server
Refactor Tracker so that generating a lookup table for rendering
is done in Tracker instead of App. App now passes the full logs
list (for the current month) to Tracker. This helps with DRY too.

Fix and simplify handleToggleCell logic to work with json-server,
which is now possible because Tracker handles its own render
preprocessing.