fix: Change Base.find_by_id behavior to disregard status field
find_by_id will always return the resource with given id, even if
the resource status is 'deleted'. 'status' should only matter in
GET handlers, therefore should only be handled in GET handlers.

This also fixes a bug where a resource marked as 'deleted' will not
be able to be subsequently marked as 'active' again.